fix(port): #142 talking to Tingle crashes — GetFuserId u64 index wild-deref
GetFuserId is a multi-return function packing the fuser id (low 32 bits) and
fuser text id (high 32 bits). Upstream zeldaret/tmc declares it u32, so the
simple callers `gSave.kinstones.fuserProgress[GetFuserId(this)]`
(tingleSiblings.c:155, din.c, farore.c, nayru.c) index the 128-byte array with
just the fuser id. Port commit c8dfdb16d widened the declaration to u64; on
64-bit PC the full u64 return is then used as the array index, so the non-zero
text id in the high half pushes the subscript ~2.7 TB out of range -> SIGSEGV.
On GBA the index is 32-bit regardless, so it never showed there.

Root-caused from a Linux backtrace (build/pc bugreport): crash IP resolves to
sub_08064F28 at tingleSiblings.c:155, fault addr 0x58bd... far from the binary
base. Disassembly confirmed `movzbl (%rax,%rdx,1)` using the full 64-bit rax as
the index.

Fix restores the u32 declaration to match upstream — the four simple sites now
index with the fuser id only; GetFuserIdAndFuserTextId (kinstone menu) is
unaffected since it reinterpret-casts the function pointer to read both halves.
Verified at the instruction level: the post-fix asm does `mov %eax,%edx` so the
index is fuserId (<=255), never the high half.

+++ include/asm.h
@@ -12,7 +12,17 @@ extern void SetTile(u32 tileIndex, u32 tilePos, u32 layer);
 extern void UpdateScrollVram(void);
 extern u32 sub_080B1BA4(u32, u32, u32);
 extern void LoadResourceAsync(const void* src, void* dest, u32 size);
-extern u64 GetFuserId(struct Entity_*);
+/* Issue #142 (talking to Tingle crashes): GetFuserId packs two results — the
+ * fuser id (low 32) and the fuser text id (high 32). On GBA it's declared u32
+ * (upstream zeldaret/tmc), so the simple callers `gSave.kinstones.fuserProgress[
+ * GetFuserId(this)]` (tingleSiblings/din/farore/nayru) index with just the
+ * fuser id. A port commit (7ccca6769) widened this to u64; on 64-bit PC the
+ * un-truncated index then includes the non-zero high half (text id), so the
+ * array subscript lands ~2.7 TB out of range and SIGSEGVs. On GBA the index is
+ * 32-bit regardless, so it never showed there. Restore the u32 declaration to
+ * match upstream — callers needing both halves use GetFuserIdAndFuserTextId,
+ * which reinterpret-casts the function pointer and is unaffected by this type. */
+extern u32 GetFuserId(struct Entity_*);
 #define GetFuserIdAndFuserTextId(ent) ((union SplitDWord)(*(MultiReturnTypeSingleEntityArg)(&GetFuserId))(ent))
 extern u32 CheckPlayerInRegion(u32 centerX, u32 centerY, u32 radiusX, u32 radiusY);
 extern u32 GravityUpdate(struct Entity_* entity, u32 gravity);
